Moved constraints to different file.
